Conflicts:
	applications/plugins/Compliant/mapping/ContactMapping.h
	applications/plugins/Compliant/mapping/RigidJointMultiMapping.h
	applications/plugins/Flexible/deformationMapping/BaseDeformationMultiMapping.inl
	applications/plugins/image/CMakeLists.txt
	extlibs/qwt-6.0.1/src/CMakeLists.txt
	framework/sofa/core/MultiMapping.h
	framework/sofa/core/MultiMapping.inl

Former-commit-id: e8fe2e5edaad57eda0ba50c8828b9bf76957a7bd

+++ modules/sofa/component/topology/MeshTopology.cpp
@@ -80,7 +80,7 @@ void MeshTopology::EdgeUpdate::updateFromVolume()
     SeqEdges& seqEdges = *topology->seqEdges.beginEdit();
     seqEdges.clear();
     std::map<Edge,unsigned int> edgeMap;
-    int edgeIndex;
+    unsigned int edgeIndex;
 
     const SeqTetrahedra& tetrahedra = topology->getTetrahedra(); // do not use seqTetrahedra directly as it might not be up-to-date
     const unsigned int edgesInTetrahedronArray[6][2]= {{0,1},{0,2},{0,3},{1,2},{1,3},{2,3}};
@@ -104,7 +104,7 @@ void MeshTopology::EdgeUpdate::updateFromVolume()
             if (ite==edgeMap.end())
             {
                 // edge not in edgeMap so create a new one
-                edgeIndex=seqEdges.size();
+                edgeIndex=(unsigned int)seqEdges.size();
                 edgeMap[e]=edgeIndex;
                 seqEdges.push_back(e);
             }
@@ -144,7 +144,7 @@ void MeshTopology::EdgeUpdate::updateFromVolume()
             if (ite==edgeMap.end())
             {
                 // edge not in edgeMap so create a new one
-                edgeIndex=seqEdges.size();
+                edgeIndex=(unsigned int)seqEdges.size();
                 edgeMap[e]=edgeIndex;
                 seqEdges.push_back(e);
 
@@ -168,7 +168,7 @@ void MeshTopology::EdgeUpdate::updateFromSurface()
     typedef MeshTopology::Quad         Quad;
 
     std::map<Edge,unsigned int> edgeMap;
-    int edgeIndex;
+    unsigned int edgeIndex;
     SeqEdges& seqEdges = *topology->seqEdges.beginEdit();
     seqEdges.clear();
     const SeqTriangles& triangles = topology->getTriangles(); // do not use seqTriangles directly as it might not be up-to-date
@@ -191,7 +191,7 @@ void MeshTopology::EdgeUpdate::updateFromSurface()
             if (ite==edgeMap.end())
             {
                 // edge not in edgeMap so create a new one
-                edgeIndex=seqEdges.size();
+                edgeIndex=(unsigned int)seqEdges.size();
                 edgeMap[e]=edgeIndex;
                 seqEdges.push_back(e);
             }
@@ -223,7 +223,7 @@ void MeshTopology::EdgeUpdate::updateFromSurface()
             if (ite==edgeMap.end())
             {
                 // edge not in edgeMap so create a new one
-                edgeIndex=seqEdges.size();
+                edgeIndex=(unsigned int)seqEdges.size();
                 edgeMap[e]=edgeIndex;
                 seqEdges.push_back(e);
             }
@@ -259,7 +259,7 @@ void MeshTopology::TriangleUpdate::update()
     std::map<Triangle,unsigned int> triangleMap;
     std::map<Triangle,unsigned int>::iterator itt;
     Triangle tr;
-    int triangleIndex;
+    unsigned int triangleIndex;
     unsigned int v[3],val;
     /// create the m_edge array at the same time than it fills the m_trianglesInTetrahedron array
     for (unsigned int i = 0; i < tetrahedra.size(); ++i)
@@ -287,7 +287,7 @@ void MeshTopology::TriangleUpdate::update()
             if (itt==triangleMap.end())
             {
                 // edge not in edgeMap so create a new one
-                triangleIndex=seqTriangles.size();
+                triangleIndex=(unsigned int)seqTriangles.size();
                 tr=Triangle(v[0],v[1],v[2]);
                 triangleMap[tr]=triangleIndex;
                 seqTriangles.push_back(tr);
@@ -329,7 +329,7 @@ void MeshTopology::QuadUpdate::update()
     std::map<Quad,unsigned int>::iterator itt;
     Quad qu;
     unsigned int v[4],val;
-    int quadIndex;
+    unsigned int quadIndex;
     /// create the m_edge array at the same time than it fills the m_edgesInHexahedron array
     for (unsigned int i = 0; i < hexahedra.size(); ++i)
     {
@@ -351,7 +351,7 @@ void MeshTopology::QuadUpdate::update()
         if (itt==quadMap.end())
         {
             // quad not in edgeMap so create a new one
-            quadIndex=seqQuads.size();
+            quadIndex=(unsigned int)seqQuads.size();
             quadMap[qu]=quadIndex;
             qu=Quad(v[0],v[1],v[2],v[3]);
             quadMap[qu]=quadIndex;
@@ -380,7 +380,7 @@ void MeshTopology::QuadUpdate::update()
         if (itt==quadMap.end())
         {
             // quad not in edgeMap so create a new one
-            quadIndex=seqQuads.size();
+            quadIndex=(unsigned int)seqQuads.size();
             quadMap[qu]=quadIndex;
             qu=Quad(v[0],v[1],v[2],v[3]);
             quadMap[qu]=quadIndex;
@@ -409,7 +409,7 @@ void MeshTopology::QuadUpdate::update()
         if (itt==quadMap.end())
         {
             // quad not in edgeMap so create a new one
-            quadIndex=seqQuads.size();
+            quadIndex=(unsigned int)seqQuads.size();
             quadMap[qu]=quadIndex;
             qu=Quad(v[0],v[1],v[2],v[3]);
             quadMap[qu]=quadIndex;
@@ -437,7 +437,7 @@ void MeshTopology::QuadUpdate::update()
         if (itt==quadMap.end())
         {
             // quad not in edgeMap so create a new one
-            quadIndex=topology->seqQuads.getValue().size();
+            quadIndex=(unsigned int)topology->seqQuads.getValue().size();
             quadMap[qu]=quadIndex;
             qu=Quad(v[0],v[1],v[2],v[3]);
             quadMap[qu]=quadIndex;
@@ -465,7 +465,7 @@ void MeshTopology::QuadUpdate::update()
         if (itt==quadMap.end())
         {
             // quad not in edgeMap so create a new one
-            quadIndex=seqQuads.size();
+            quadIndex=(unsigned int)seqQuads.size();
             quadMap[qu]=quadIndex;
             qu=Quad(v[0],v[1],v[2],v[3]);
             quadMap[qu]=quadIndex;
@@ -493,7 +493,7 @@ void MeshTopology::QuadUpdate::update()
         if (itt==quadMap.end())
         {
             // quad not in edgeMap so create a new one
-            quadIndex=seqQuads.size();
+            quadIndex=(unsigned int)seqQuads.size();
             quadMap[qu]=quadIndex;
             qu=Quad(v[0],v[1],v[2],v[3]);
             quadMap[qu]=quadIndex;
@@ -670,8 +670,8 @@ void MeshTopology::addPoint(double px, double py, double pz)
 {
     seqPoints.beginEdit()->push_back(defaulttype::Vec<3,SReal>((SReal)px, (SReal)py, (SReal)pz));
     seqPoints.endEdit();
-    if (seqPoints.getValue().size() > (unsigned)nbPoints)
-        nbPoints = seqPoints.getValue().size();
+    if (seqPoints.getValue().size() > (size_t)nbPoints)
+        nbPoints = (int)seqPoints.getValue().size();
 }
 
 void MeshTopology::addEdge( int a, int b )
@@ -733,8 +733,8 @@ void MeshTopology::addUV(double u, double v)
 {
     seqUVs.beginEdit()->push_back(defaulttype::Vec<2,SReal>((SReal)u, (SReal)v));
     seqUVs.endEdit();
-    if (seqUVs.getValue().size() > (unsigned)nbPoints)
-        nbPoints = seqUVs.getValue().size();
+    if (seqUVs.getValue().size() > (size_t)nbPoints)
+        nbPoints = (int)seqUVs.getValue().size();
 }
 
 const MeshTopology::SeqEdges& MeshTopology::getEdges()
@@ -790,32 +790,32 @@ void MeshTopology::setNbPoints(int n)
 
 int MeshTopology::getNbEdges()
 {
-    return getEdges().size();
+    return (int)getEdges().size();
 }
 
 int MeshTopology::getNbTriangles()
 {
-    return getTriangles().size();
+    return (int)getTriangles().size();
 }
 
 int MeshTopology::getNbQuads()
 {
-    return getQuads().size();
+    return (int)getQuads().size();
 }
 
 int MeshTopology::getNbTetrahedra()
 {
-    return getTetrahedra().size();
+    return (int)getTetrahedra().size();
 }
 
 int MeshTopology::getNbHexahedra()
 {
-    return getHexahedra().size();
+    return (int)getHexahedra().size();
 }
 
 int MeshTopology::getNbUVs()
 {
-    return getUVs().size();
+    return (int)getUVs().size();
 }
 
 const MeshTopology::Edge MeshTopology::getEdge(index_type i)
@@ -2380,7 +2380,7 @@ unsigned int MeshTopology::getNumberOfConnectedComponent()
     while (elemAll.size() < nbr)
     {
         std::sort(elemAll.begin(), elemAll.end());
-        unsigned int other_ID = elemAll.size();
+        unsigned int other_ID = (unsigned int)elemAll.size();
 
         for (unsigned int i = 0; i<elemAll.size(); ++i)
             if (elemAll[i] != i)
@@ -2451,7 +2451,7 @@ const sofa::helper::vector <unsigned int> MeshTopology::getConnectedElement(unsi
         }
 
         // cpt for connexity
-        cpt +=elemPreviousFront.size();
+        cpt += (unsigned int)elemPreviousFront.size();
 
         if (elemPreviousFront.empty())
         {
